Split walk-in onboarding shell and stages
